+++ lib/metrics/prometheus.go
@@ -21,9 +21,12 @@
 package metrics
 
 import (
+	"fmt"
 	"net/http"
 	"strings"
 	"sync"
+	"sync/atomic"
+	"time"
 
 	"github.com/Jeffail/benthos/lib/log"
 	"github.com/prometheus/client_golang/prometheus"
@@ -37,10 +40,17 @@ func init() {
 	constructors[TypePrometheus] = typeSpec{
 		constructor: NewPrometheus,
 		description: `
-Host endpoints for Prometheus scraping. The field ` + "`push_url`" + ` is
-optional and when set will trigger a push of metrics once Benthos shuts down.
-This is useful for when Benthos instances are short lived. Do not include the
-"/metrics/jobs/..." path to the push URL.`,
+Host endpoints for Prometheus scraping.
+
+### Push Gateway
+
+The field ` + "`push_url`" + ` is optional and when set will trigger a push of
+metrics to a [Prometheus Push Gateway](https://prometheus.io/docs/instrumenting/pushing/)
+once Benthos shuts down. It is also possible to specify a
+` + "`push_interval`" + ` which results in periodic pushes.
+
+The Push Gateway This is useful for when Benthos instances are short lived. Do
+not include the "/metrics/jobs/..." path in the push URL.`,
 	}
 }
 
@@ -48,13 +58,15 @@ This is useful for when Benthos instances are short lived. Do not include the
 
 // PrometheusConfig is config for the Prometheus metrics type.
 type PrometheusConfig struct {
-	PushURL string `json:"push_url" yaml:"push_url"`
+	PushURL      string `json:"push_url" yaml:"push_url"`
+	PushInterval string `json:"push_interval" yaml:"push_interval"`
 }
 
 // NewPrometheusConfig creates an PrometheusConfig struct with default values.
 func NewPrometheusConfig() PrometheusConfig {
 	return PrometheusConfig{
-		PushURL: "",
+		PushURL:      "",
+		PushInterval: "",
 	}
 }
 
@@ -151,6 +163,10 @@ func (p *PromGaugeVec) With(labelValues ...string) StatGauge {
 // Prometheus is a stats object with capability to hold internal stats as a JSON
 // endpoint.
 type Prometheus struct {
+	log        log.Modular
+	closedChan chan struct{}
+	running    int32
+
 	config PrometheusConfig
 	prefix string
 
@@ -164,15 +180,37 @@ type Prometheus struct {
 // NewPrometheus creates and returns a new Prometheus object.
 func NewPrometheus(config Config, opts ...func(Type)) (Type, error) {
 	p := &Prometheus{
-		config:   config.Prometheus,
-		prefix:   toPromName(config.Prefix),
-		counters: map[string]*prometheus.CounterVec{},
-		gauges:   map[string]*prometheus.GaugeVec{},
-		timers:   map[string]*prometheus.SummaryVec{},
+		log:        log.Noop(),
+		running:    1,
+		closedChan: make(chan struct{}),
+		config:     config.Prometheus,
+		prefix:     toPromName(config.Prefix),
+		counters:   map[string]*prometheus.CounterVec{},
+		gauges:     map[string]*prometheus.GaugeVec{},
+		timers:     map[string]*prometheus.SummaryVec{},
 	}
 
 	for _, opt := range opts {
 		opt(p)
+	}
+
+	if len(p.config.PushURL) > 0 && len(p.config.PushInterval) > 0 {
+		interval, err := time.ParseDuration(p.config.PushInterval)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse push interval: %v", err)
+		}
+		go func() {
+			for {
+				select {
+				case <-p.closedChan:
+					return
+				case <-time.After(interval):
+					if err = push.New(p.config.PushURL, "benthos_push").Gatherer(prometheus.DefaultGatherer).Push(); err != nil {
+						p.log.Errorf("Failed to push metrics: %v\n", err)
+					}
+				}
+			}
+		}()
 	}
 
 	return p, nil
@@ -347,11 +385,15 @@ func (p *Prometheus) GetGaugeVec(path string, labelNames []string) StatGaugeVec 
 
 // SetLogger does nothing.
 func (p *Prometheus) SetLogger(log log.Modular) {
+	p.log = log
 }
 
 // Close stops the Prometheus object from aggregating metrics and cleans up
 // resources.
 func (p *Prometheus) Close() error {
+	if atomic.CompareAndSwapInt32(&p.running, 1, 0) {
+		close(p.closedChan)
+	}
 	if len(p.config.PushURL) > 0 {
 		return push.New(p.config.PushURL, "benthos_push").Gatherer(prometheus.DefaultGatherer).Push()
 	}